Move type stuff into separate files and make types inherit namespaces
