+++ src/constructions/fallthrough.hpp
@@ -6,7 +6,6 @@
 #include <llvm/IR/IRBuilder.h>
 #include <string>
 #include "../types/type.hpp"
-#include "../types/builtins.hpp"
 
 namespace AST {
 	class FallthroughExpression: public AST::Expression {
