+++ src/constructions/function_call.cpp
@@ -24,7 +24,6 @@
 #include "../parsing.hpp"
 #include "codeblock.hpp"
 #include "../mangling.hpp"
-#include "../types/builtins.hpp"
 
 llvm::Value* AST::FunctionCall::codegen(llvm::IRBuilder<> *builder, std::string name) {
 	/*llvm::Function *F = function->generated;
