+++ src/constructions/function_call.hpp
@@ -14,7 +14,6 @@
 #include <llvm/IR/Type.h>
 #include <llvm/IR/Instructions.h>
 #include <llvm/IR/Function.h>
-#include "../types/builtins.hpp"
 
 #include "function_def.hpp"
