+++ src/constructions/numeric_literal.cpp
@@ -1,6 +1,5 @@
 #include "numeric_literal.hpp"
 #include "../parsing.hpp"
-#include "../types/builtins.hpp"
 #include "construction_parse.hpp"
 #include "../macros.hpp"
